packages/server/requirements.txt
django
Pillow
djangorestframework
django-extensions